server/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
requests==2.31.0

# LLM Providers (install only what you need)
groq==0.4.2              # FREE tier - recommended for testing
openai==1.12.0           # Optional - paid
anthropic==0.18.1        # Optional - paid
